+++ src/ExpenseForm.jsx
@@ -1,0 +1,44 @@
+import { useState } from 'react';
+
+function ExpenseForm({ onAddExpense }) {
+  const [name, setName] = useState('');
+  const [amount, setAmount] = useState('');
+  const [category, setCategory] = useState('Food');
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (!name || !amount) return;
+
+    onAddExpense({ name, amount: Number(amount), category });
+    setName('');
+    setAmount('');
+    setCategory('Food');
+  };
+
+  return (
+    <form onSubmit={handleSubmit}>
+      <h3>Add Expense</h3>
+      <input
+        type="text"
+        placeholder="Expense name"
+        value={name}
+        onChange={(e) => setName(e.target.value)}
+      />
+      <input
+        type="number"
+        placeholder="Amount"
+        value={amount}
+        onChange={(e) => setAmount(e.target.value)}
+      />
+      <select value={category} onChange={(e) => setCategory(e.target.value)}>
+        <option>Food</option>
+        <option>Utilities</option>
+        <option>Transport</option>
+        <option>Entertainment</option>
+      </select>
+      <button type="submit">Add Expense</button>
+    </form>
+  );
+}
+
+export default ExpenseForm;
