+++ src/Header.jsx
@@ -1,0 +1,10 @@
+function Header() {
+    return (
+      <header>
+        <h1>Expense Tracker</h1>
+      </header>
+    );
+  }
+  
+  export default Header;
+
