--- src/Header.new.jsx
@@ -1,10 +1,0 @@
-function Header() {
-    return (
-      <header>
-        <h1>Expense Tracker</h1>
-      </header>
-    );
-  }
-  
-  export default Header;
-
